requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pytest==8.4.2
selenium~=4.37.0
webdriver-manager==4.0.2
allure-pytest==2.15.3